app/pyproject.toml
[project]
name = "telescope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-instrumentation>=0.55b1",
    "opentelemetry-sdk>=1.34.1",
    "uvicorn[standard]>=0.34.3",
]
